+++ lib/src/image_viewer.dart
@@ -6,8 +6,9 @@ import 'image_body.dart';
 class ImageViewer extends ImageBody {
   ImageViewer({
     super.onTap,
-    required final String image,
-    super.aliveName,
+    super.htmlImage,
+    super.assetImage,
+    super.alive,
     super.width,
     super.height,
     super.onErrorWidget,
@@ -25,8 +26,8 @@ class ImageViewer extends ImageBody {
     super.duration,
     super.maxSize,
   }) :
+  assert( htmlImage == null || assetImage == null, "Only one image must be provided" ),
   super(
-    onLoadingImage: image,
     readOnly      : true,
     filterOpacity : 0,
     viewerBlur    : false,
@@ -34,8 +35,9 @@ class ImageViewer extends ImageBody {
 
   factory ImageViewer.square({
     final Future<void> Function() ? onTap,
-    required final  String           image,
-    final String                  ? aliveName,
+    final String                  ? htmlImage,
+    final String                  ? assetImage,
+    final bool                    ? alive,
     final double                  ? dimension,
     final Widget                  ? emptyWidget,
     final Widget                  ? filledWidget,
@@ -54,8 +56,9 @@ class ImageViewer extends ImageBody {
   }) =>
   ImageViewer(
     onTap           : onTap,
-    image           : image,
-    aliveName       : aliveName,
+    htmlImage       : htmlImage,
+    assetImage      : assetImage,
+    alive           : alive,
     width           : dimension,
     height          : dimension,
     onErrorWidget   : onErrorWidget,
@@ -75,8 +78,9 @@ class ImageViewer extends ImageBody {
 
   factory ImageViewer.circle({
     final Future<void> Function() ? onTap,
-    required final String           image,
-    final String                  ? aliveName,
+    final String                  ? htmlImage,
+    final String                  ? assetImage,
+    final bool                    ? alive,
     final double                  ? dimension,
     final Widget                  ? emptyWidget,
     final Widget                  ? filledWidget,
@@ -94,8 +98,9 @@ class ImageViewer extends ImageBody {
   }) =>
   ImageViewer(
     onTap           : onTap,
-    image           : image,
-    aliveName       : aliveName,
+    htmlImage       : htmlImage,
+    assetImage      : assetImage,
+    alive       : alive,
     width           : dimension,
     height          : dimension,
     onErrorWidget   : onErrorWidget,
@@ -114,8 +119,9 @@ class ImageViewer extends ImageBody {
 
   factory ImageViewer.expand({
     final Future<void> Function() ? onTap,
-    required final String           image,
-    final String                  ? aliveName,
+    final String                  ? htmlImage,
+    final String                  ? assetImage,
+    final bool                    ? alive,
     final Widget                  ? emptyWidget,
     final Widget                  ? filledWidget,
     final Widget                  ? onErrorWidget,
@@ -133,8 +139,9 @@ class ImageViewer extends ImageBody {
   }) =>
   ImageViewer(
     onTap           : onTap,
-    image           : image,
-    aliveName       : aliveName,
+    htmlImage       : htmlImage,
+    assetImage      : assetImage,
+    alive           : alive,
     width           : double.infinity,
     height          : double.infinity,
     onErrorWidget   : onErrorWidget,